requirements.txt
opencv-contrib-python==3.3.0.9
opencv-python==3.3.0.9
requests==2.18.4
h5py==2.8.0
Keras==2.1.6
numpy==1.14.3
Pillow==5.1.0
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1

